Added basic UI for company list and company details, all projects are integrated to work as a single solution

+++ Core/AlexDemo.CustomerHub.Core.Application/UseCases/Customer/Company/Handlers/Commands/DeleteCompanyCommandHandler.cs
@@ -1,10 +1,12 @@
 ﻿using AlexDemo.CustomerHub.Core.Application.Contracts.Persistence.Customer;
+using AlexDemo.CustomerHub.Core.Application.Enums;
+using AlexDemo.CustomerHub.Core.Application.Responses;
 using AlexDemo.CustomerHub.Core.Application.UseCases.Customer.Company.Actions.Commands;
 using AlexDemo.CustomerHub.Core.Enums;
 
 namespace AlexDemo.CustomerHub.Core.Application.UseCases.Customer.Company.Handlers.Commands
 {
-    public class DeleteCompanyCommandHandler : IRequestHandler<DeleteCompanyCommand, Unit>
+    public class DeleteCompanyCommandHandler : IRequestHandler<DeleteCompanyCommand, BaseCommandResponse>
     {
         private readonly ICompanyRepository _companyRepository;
         private readonly IMapper _mapper;
@@ -15,7 +17,7 @@ namespace AlexDemo.CustomerHub.Core.Application.UseCases.Customer.Company.Handle
             _mapper = mapper;
         }
 
-        public async Task<Unit> Handle(DeleteCompanyCommand request, CancellationToken cancellationToken)
+        public async Task<BaseCommandResponse> Handle(DeleteCompanyCommand request, CancellationToken cancellationToken)
         {
             if (request.Id <= 0)
             {
@@ -27,16 +29,36 @@ namespace AlexDemo.CustomerHub.Core.Application.UseCases.Customer.Company.Handle
                 throw new ArgumentException(nameof(request.Actor));
             }
 
-            var company = await _companyRepository.GetById(request.Id);
-            if (company.Status == Status.Deleted)
+            var response = new BaseModifyCommandResponse<int>
             {
-                // there's no point to delete already deleted entity
-                throw new ApplicationException("Entity already deleted");
+                Id = request.Id
+            };
+
+
+            var company = await _companyRepository.GetById(request.Id);
+            if (company == null || company.Status == Status.Deleted)
+            {
+                // an alternative approach not to throw exceptions but to work with response types
+                response.IsSuccessful = false;
+                response.Message = "Delete Company Failed";
+                response.Data =
+                [
+                    new ResponseMessageModel
+                    {
+                        ResponseType = ResponseType.ValidationError,
+                        Message = "already deleted"
+                    }
+
+                ];
+
+                return response;
             }
 
             await _companyRepository.Delete(company);
 
-            return Unit.Value;
+            response.IsSuccessful = true;
+            response.Message = "Company deleted";
+            return response;
         }
     }
 }
